requirements.txt
easydict
tqdm
PyYAML
pandas
loguru
scikit-learn
kneed
xgboost
lightgbm
mlflow
PyYAML
pytest
mkdocs-material
tox
streamlit
# -e .